Version controlled controlled vocabularies #58
- created script to download controlled vocabularies from ODM2 API and
store as version controlled json files in
src/odm2_postgres_api/cv_definitions
- load vocabularies on app startup by reading in json files and insert
if not already in database

this improves startup significantly, going from ~40 secs each startup
(probably a lot slower on slow internet) to:

- 6-7 seconds first time startup
- 1-2 seconds after first time

TODO: consider to improve function to write cv's to do async writes to
improve write speed further?
