+++ src/odm2_postgres_api/controlled_vocabularies/cv_definitions/README.md
@@ -1,0 +1,5 @@
+# ODM2 controlled vocabularies
+
+These json files are controlled vocabularies defined in external API http://vocabulary.odm2.org/. The files are downloaded by script [download_cvs.py](../download_cvs.py) and should not be edited manually. Any edits would be overwritten when running this script.
+
+The procedure is to manually run script when we want to update and check into git.
