+++ src/odm2_postgres_api/controlled_vocabularies/cv_definitions/cv_directivetype.json
@@ -1,0 +1,45 @@
+{
+  "meta": {
+    "limit": 0,
+    "offset": 0,
+    "total_count": 3
+  },
+  "objects": [
+    {
+      "category": "",
+      "definition": "A sampling event conducted in the field during which instruments may be deployed and during which samples may be collected. Field campaigns typically have a focus such as characterizing a particular environment, quantifying a particular phenomenon, answering a particular research question, etc. and may last for hours, days, weeks, months, or even longer.",
+      "name": "Field campaign",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/directivetype/fieldCampaign/",
+      "term": "fieldCampaign",
+      "vocabulary_id": 294,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "Environmental monitoring that is conducted according to a formal plan that may reflect the overall objectives of an organization, references specific strategies that help deliver the objectives and details of specific projects or tasks, and that contains a listing of what is being monitored, how that monitoring is taking place, and the time-scale over which monitoring should take place.",
+      "name": "Monitoring program",
+      "note": "",
+      "provenance": "Definition adapted from Wikipedia.  See http://en.wikipedia.org/wiki/Environmental_monitoring.",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/directivetype/monitoringProgram/",
+      "term": "monitoringProgram",
+      "vocabulary_id": 295,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "A collaborative enterprise, involving research or design, the is carefully planned to achieve a particular aim.",
+      "name": "Project",
+      "note": "",
+      "provenance": "Definition adapted from Wikipedia.  See http://en.wikipedia.org/wiki/Project.",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/directivetype/project/",
+      "term": "project",
+      "vocabulary_id": 296,
+      "vocabulary_status": "Current"
+    }
+  ]
+}
