+++ src/odm2_postgres_api/controlled_vocabularies/cv_definitions/cv_propertydatatype.json
@@ -1,0 +1,69 @@
+{
+  "meta": {
+    "limit": 0,
+    "offset": 0,
+    "total_count": 5
+  },
+  "objects": [
+    {
+      "category": "",
+      "definition": "A boolean type is typically a logical type that can be either \"true\" or \"false\".",
+      "name": "Boolean",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/propertydatatype/boolean/",
+      "term": "boolean",
+      "vocabulary_id": 1515,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "A predefined list of strings that a user can select from.",
+      "name": "Controlled list",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/propertydatatype/controlledList/",
+      "term": "controlledList",
+      "vocabulary_id": 1526,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "A floating-point number represents a limited-precision rational number that may have a fractional part. ",
+      "name": "Floating point number",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/propertydatatype/floatingPointNumber/",
+      "term": "floatingPointNumber",
+      "vocabulary_id": 1516,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "An integer data type can hold a whole number, but no fraction. Integers may be either signed (allowing negative values) or unsigned (nonnegative values only). ",
+      "name": "Integer",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/propertydatatype/integer/",
+      "term": "integer",
+      "vocabulary_id": 1517,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "An array of characters including letters, digits, punctuation marks, symbols, etc.",
+      "name": "String",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/propertydatatype/string/",
+      "term": "string",
+      "vocabulary_id": 1518,
+      "vocabulary_status": "Current"
+    }
+  ]
+}
