+++ src/odm2_postgres_api/controlled_vocabularies/cv_definitions/cv_status.json
@@ -1,0 +1,57 @@
+{
+  "meta": {
+    "limit": 0,
+    "offset": 0,
+    "total_count": 4
+  },
+  "objects": [
+    {
+      "category": "",
+      "definition": "Data collection is complete. No new data values will be added.",
+      "name": "Complete",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/status/complete/",
+      "term": "complete",
+      "vocabulary_id": 1103,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "Data collection is ongoing.  New data values will be added periodically.",
+      "name": "Ongoing",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/status/ongoing/",
+      "term": "ongoing",
+      "vocabulary_id": 1104,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "Data collection is planned. Resulting data values do not yet exist, but will eventually.",
+      "name": "Planned",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/status/planned/",
+      "term": "planned",
+      "vocabulary_id": 1105,
+      "vocabulary_status": "Current"
+    },
+    {
+      "category": "",
+      "definition": "The status of data collection is unknown.",
+      "name": "Unknown",
+      "note": "",
+      "provenance": "",
+      "provenance_uri": "",
+      "resource_uri": "/api/v1/status/unknown/",
+      "term": "unknown",
+      "vocabulary_id": 1106,
+      "vocabulary_status": "Current"
+    }
+  ]
+}
